chore(Pendo): move sdk loader to different file
